pyproject.toml
[tool.poetry]
name = "pest"
version = "0.1.0"
description = "Preprocessing Engine for Spherinator Training"
authors = [
    "Bernd Doser <bernd.doser@h-its.org>",
    "Sebastian T. Gomez <sebastian.trujillogomez@h-its.org>",
]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "pest", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
h5py = "^3.11.0"
requests = "^2.32.3"
illustris-python = {git = "https://github.com/illustristng/illustris_python.git"}
matplotlib = "^3.9.2"
scipy = "^1.14.1"
pyyaml = "^6.0.2"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
